update booking page
