+++ src/App.css
@@ -1,0 +1,3 @@
+.pointer-hover:hover {
+    cursor: pointer;
+}
