+++ src/components/doctor/doctorProfile.jsx
@@ -1,4 +1,4 @@
-export default function DoctorProfile({doctor}) {
+export default function DoctorProfile({ doctor }) {
   return (
     <div className="widget-profile pro-widget-content">
       <div className="profile-info-widget">
@@ -8,7 +8,8 @@ export default function DoctorProfile({doctor}) {
         <div className="profile-det-info">
           <h3>{doctor.fullname}</h3>
           <div className="patient-details">
-            <h5 className="mb-0">{doctor.department_name}</h5>
+            <h5 className="mb-0">Specialty: {doctor.specialize}</h5>
+            <h5 className="mb-0">In {doctor.department_name ?  doctor.department_name : "..."} Department</h5>
           </div>
         </div>
       </div>
